Moved helper functions to separate utils file
