+++ server/utils.js
@@ -1,3 +1,5 @@
+/* SERVER SIDE */
+
 const checkConditionMiddleware = (conditionFuncs) => (req, res, next) => {
   const finalCondition = conditionFuncs.map(conditionFunc =>
     conditionFunc(req)).reduce((prev, curr) => prev && curr, true);
@@ -10,6 +12,7 @@ const checkConditionMiddleware = (conditionFuncs) => (req, res, next) => {
 };
 
 const adminConditionFunc = (req) => req.user && req.user.isAdmin;
+
 
 module.exports = {
   checkConditionMiddleware,
